local-management/src/requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
Django==2.2
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==2.2.0
django-crispy-forms==1.9.0
django-rest-knox==4.1.0
django-timezone-field==4.0
django-widget-tweaks==1.4.5
djangorestframework==3.8.2
httpie==2.0.0
idna==2.9
kombu==4.6.7
numpy==1.18.4
Pillow==7.0.0
pycparser==2.20
Pygments==2.6.1
pyproj==2.6.1.post1
python-crontab==2.4.2
python-dateutil==2.8.1
pytz==2018.4
redis==3.4.1
requests==2.23.0
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
vine==1.3.0
